Move test into separate file
